Map measure updates (#92)
* commiting changes

* side-bar update

* map measure minor updates

* completed map measure updates

* minor updates to side bar

* minor changes

* making station profile required

* sidebar updates and station creation form updates

* minor updates to station metadata
